scanner-worker/requirements.txt
minio==7.2.4
redis==5.0.3
psycopg2-binary==2.9.9
pyclamd==0.4.0